+++ milkman-explore/src/main/java/milkman/plugin/explore/JqlQueryAspect.java
@@ -1,9 +1,13 @@
 package milkman.plugin.explore;
 
+import java.util.LinkedList;
+import java.util.List;
+
 import lombok.Data;
 import milkman.domain.RequestAspect;
 
 @Data
 public class JqlQueryAspect extends RequestAspect {
 	private String query = "";
+	private List<String> queryHistory = new LinkedList<>();
 }